python/requirements.txt
numpy
pandas
scipy
matplotlib
requests
Pillow
scikit-learn
tensorflow
